work on rewriting time rift shader effect
