+++ Assets/Materials/Time Rift.mat.meta
@@ -1,7 +1,8 @@
 fileFormatVersion: 2
-guid: cd4de1267569a6d4eb31c3f60600033b
-DefaultImporter:
+guid: 3da5687dc6614fa4e8b05340252bf016
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 2100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
