+++ Assets/Models/sphere inverted normals.fbx
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:dac2ad7296daad12e509facd3b2d712f7194cb9bf54e1acd56590539257b3a3a
+size 31820
