+++ Assets/Scripts/Features/Time Rift/Time Rift Sphere.shader.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 80e8ed300b613e842a142b2d6e2e03d8
+ShaderImporter:
+  externalObjects: {}
+  defaultTextures: []
+  nonModifiableTextures: []
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
