--- Assets/Scripts/Features/Time Rift/TimeBubbleUnlit.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 70634a9925f740d7a653d518f5342f9b
-timeCreated: 1696847227
